New Entity: Shade
The Shade, although not my favorite, I did have fun coding in cause it utilizes the player's skin (client-side) and overlays a transparent texture of noise I quickly generated. The mob is slow but packs a punch and although it can be killed through normal methods it would take forever, however with this new entity comes a new Tag simply called Torches which contains all current torch-types and is used for the Shade for once you right-click the shade with one of the torches in the tag, they instantly keel over and die!